refactor(datatool): split monolithic util.rs into per-subcommand modules
